cloud_functions/api_router/requirements.txt
functions-framework==3.*
requests
firebase-admin
google-cloud-firestore
google-cloud-functions
google-cloud-run
firebase-functions
beartype
